Moved funcs to a new file
